ai-fastapi/requirements.txt
torch>=2.5.1
ultralytics
sam2
open_clip_torch
fastapi
uvicorn
python-multipart
numpy
opencv-python
